Changed experiment files in preparation for testing.
Changed the experiment files for the n = 10 robot case to use the intended parameters for when we conduct our experiment (i.e. the avoidance radius of robots, arena size). Also added "light_x" and "light_y" parameters that contain the location of the light beacon so that we can access them in the robot controllers. Originally wanted to opt for a more elegant solution using the magnitude of the light sensed from the light sensors, but unfortunately this reaches its maximum value, 1, while still being a fair distance from the light.

Also deleted the old named motor failure experiment file, and extended the arena size slightly so that the robots don't mistakenly run into walls when approaching the beacon.