Each experiment now terminates in a finite amount of time and produces output.
Worked with Germaine to add output to the controller in the manner we'd expect - last steps before being able to begin tests are to extract the output to a file(which we can then parse further once we finish our tests) and to create experiment files for the n=30,50,etc. cases.